Add getTweetsBeforeDate() method
Takes a list of tweets and returns a sublist of tweets which were created prior to the Date argument.
I was going to use a faster version with addAll(tweets.subList()) at the index where the first tweet that occurs before the Date argument is found, but it returned the wrong number of tweets (by one), so I stuck with the slower method. With the limits of the Twitter API, casual use of this application shouldn't see more than a few seconds of lag from this.